fault code mancanta su tabeventi,mettere token 0 nel caso di position con 0 token, extrainfo rrn riporta sendRT anzichè closePayment

+++ src/main/java/com/nexigroup/pagopa/cruscotto/sert/service/dto/PositionPaymentDTO.java
@@ -19,7 +19,7 @@ public class PositionPaymentDTO implements Serializable {
     @JsonProperty("position-info")
     private PositionPaymentInfoDTO positionInfo;
 
-    private Long tokens;
+    private Integer tokens;
 
     @JsonProperty("all-tokens")
     private List<String> allTokens;
